fix: allow browser storage sync across all modes
Problem:
Because the Vuex store does not rely on the browser extension storage
API and instead uses the Web Storage API, data is not persistent across
normal and private browsing sessions.

Solution:
Using a custom Vuex plugin, a subcription to the Vuex store is made and
for each new mutation, the extension's storage is set with the new state.
When a private session is started and it tries to access the Vuex store,
the private session's store is replaced with the up-to-date extension's
storage data, effectively syncing all Vuex store and extension storage
across all browsing modes.