feat: implement window settings handling
What's New:
The original RapidTabOpener had the option to either open tabs in a no-
rmal or private window. This feature implements this in the new version,
but it also allows users to choose whether or not they want the group
to be opened in the current window or a new one. Just a little more
power to the people.